## ✅ Document Processor Refactoring Complete
I have successfully split the large `document_processor.go` file (originally **1540 lines**) into focused, maintainable modules:

### **Files Created:**

1. **`document_processor.go`** (171 lines) - Core tool structure, init, Definition, Execute, and main orchestration
2. **`request_parsing.go`** (251 lines) - Request parsing, validation, and profile application
3. **`response_handling.go`** (159 lines) - Response formatting, file operations, and result handling
4. **`batch_processing.go`** (172 lines) - Batch processing with concurrency management
5. **`debugging.go`** (140 lines) - Debug logging, log rotation, and system info
6. **`python_integration.go`** (693 lines) - Python subprocess handling and data parsing

### **Benefits Achieved:**

✅ **Size Reduction**: Largest file now **693 lines** (down from 1540)  
✅ **Clean Separation**: Each file has a single responsibility  
✅ **Maintained Functionality**: All original features preserved  
✅ **Zero Test Failures**: All tests pass with 0 linting issues  
✅ **Build Success**: Project builds without errors  
✅ **British English**: Maintained throughout all files  

### **Code Quality Improvements:**

- **Better Maintainability**: Changes to specific functionality can be made in isolation
- **Improved Readability**: Easier to understand and navigate each concern
- **Single Responsibility**: Each file has a focused purpose
- **Preserved API**: All public methods and interfaces remain unchanged

The refactoring successfully addresses the original cleanup goal of splitting files over 800 lines whilst maintaining full functionality and code quality standards.
